requirements.txt
asgiref==3.7.2
django-cors-headers==4.3.1
dj-database-url==2.1.0
dj-static==0.0.6
Django==4.2.7
django-filter==23.5
djangorestframework==3.14.0
gunicorn==21.2.0
importlib-metadata==6.8.0
Markdown==3.5.1
psycopg2-binary==2.9.9
pytz==2023.3
Pyrebase4==4.7.1
sqlparse==0.4.4
static3==0.7.0
tzdata==2023.3
zipp==3.17.0
python-dotenv==1.0.0
django-ratelimit==4.1.0
git+https://github.com/erelsgl/fairpy.git@8beb3992f1f9269eee90c6c8a69e7ca0bbe83fd0






